Make it so locally we can use in memory sqlite for testing, but on CI mariaDB

+++ database/migrations/2025_08_09_181322_remove_use_dark_theme_from_users.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\SQLiteConnection;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
@@ -11,9 +12,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('use_dark_theme');
-        });
+        if (!DB::connection() instanceof SQLiteConnection) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('use_dark_theme');
+            });
+        }
     }
 
     /**
